Move config logic to its own package
