+++ config.yml
@@ -1,1 +1,1 @@
-CSVDataPath: pokemon.csv
+pokemon_csv_path: pokemon.csv
